Add session validation to controllers

+++ E-market_OnionMVC/Controllers/CategoriesController.cs
@@ -1,5 +1,8 @@
-﻿using E_market.Core.Application.Interfaces.Services;
+﻿using E_market.Core.Application.Helpers;
+using E_market.Core.Application.Interfaces.Services;
 using E_market.Core.Application.ViewModels.Categories;
+using E_market.Core.Application.ViewModels.Users;
+using E_market_OnionMVC.Models.Middlewares;
 using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
@@ -11,19 +14,31 @@ namespace E_market_OnionMVC.Controllers
     public class CategoriesController : Controller
     {
         private readonly ICategoryService _categoryService;
+        private readonly ValidateUserSession _validateUserSession;
 
-        public CategoriesController(ICategoryService categoryService)
+        public CategoriesController(ICategoryService categoryService, ValidateUserSession validateUserSession)
         {
             _categoryService = categoryService;
+            _validateUserSession = validateUserSession;
         }
         public async Task<IActionResult> CategoryList()
         {
+            if (!_validateUserSession.HasUser())
+            {
+                return RedirectToRoute(new { controller = "User", action = "Login" });
+            }
+
             return View(await _categoryService.GetAllViewModel());
         }
 
         [HttpPost]
         public async Task<IActionResult> Create(CategoryViewModel vm)
         {
+            if (!_validateUserSession.HasUser())
+            {
+                return RedirectToRoute(new { controller = "User", action = "Login" });
+            }
+
             if (!ModelState.IsValid)
             {
                 return View("SaveCategory", vm);
@@ -35,17 +50,32 @@ namespace E_market_OnionMVC.Controllers
 
         public IActionResult Create()
         {
+            if (!_validateUserSession.HasUser())
+            {
+                return RedirectToRoute(new { controller = "User", action = "Login" });
+            }
+
             return View("SaveCategory", new CategoryViewModel());
         }
 
         public async Task<IActionResult> Update(int id)
         {
+            if (!_validateUserSession.HasUser())
+            {
+                return RedirectToRoute(new { controller = "User", action = "Login" });
+            }
+
             return View("SaveCategory", await _categoryService.GetByIdSaveViewModel(id));
         }
 
         [HttpPost]
         public async Task<IActionResult> Update(CategoryViewModel vm)
         {
+            if (!_validateUserSession.HasUser())
+            {
+                return RedirectToRoute(new { controller = "User", action = "Login" });
+            }
+
             if (!ModelState.IsValid)
             {
                 return View("SaveCategory", vm);
@@ -57,6 +87,11 @@ namespace E_market_OnionMVC.Controllers
 
         public async Task<IActionResult> Delete(int id)
         {
+            if (!_validateUserSession.HasUser())
+            {
+                return RedirectToRoute(new { controller = "User", action = "Login" });
+            }
+
             await _categoryService.Delete(id);
             return RedirectToRoute(new { controller = "Categories", action = "CategoryList" });
         }
